refactor(api): split config.rs into config/ by responsibility
Move JwtSecret into jwt.rs, ApiConfigError into errors.rs,
env-parse helpers into env.rs, sub-config structs into sub.rs,
and ApiConfig + from_env + for_test into mod.rs. Distribute the
17 inline tests to the file that owns the type under test.
Public path nebula_api::config::* is unchanged via pub use in mod.rs.
